Refactor collision and health systems into dedicated modules
- Removed the existing collision and health system implementations, replacing them with modular structures for better organization and maintainability.
- Introduced new files for collision detection, player collision handling, and health management, enhancing clarity in the codebase.
- Updated game initialization to utilize the new modular systems, ensuring seamless integration with existing game mechanics.
- Established a clear separation of concerns within the collision and health systems, improving overall code quality and facilitating future enhancements.
