+++ internal/ecs/resources/audio.go
@@ -1,0 +1,24 @@
+package resources
+
+// Audio resource management functionality
+// This file is a placeholder for future audio resource loading and management
+// Currently, the game does not use audio resources, but this structure
+// allows for easy extension when audio features are added.
+
+// LoadAudio loads an audio resource from embedded assets
+// func (rm *ResourceManager) LoadAudio(name, path string) (*AudioResource, error) {
+//     // TODO: Implement audio loading when audio features are added
+//     return nil, nil
+// }
+
+// GetAudio retrieves a loaded audio resource
+// func (rm *ResourceManager) GetAudio(name string) (*AudioResource, bool) {
+//     // TODO: Implement audio retrieval when audio features are added
+//     return nil, false
+// }
+
+// LoadAllAudio loads all required audio resources for the game
+// func (rm *ResourceManager) LoadAllAudio() error {
+//     // TODO: Implement audio loading when audio features are added
+//     return nil
+// }
